test: handleDuplicatesInAttachments
