+++ src/app/modules/submission/email-submission/email-submission.util.ts
@@ -512,38 +512,6 @@ export const mapRouteError: MapRouteError = (error) => {
 }
 
 /**
- * Looks for duplicated filenames and changes the filename
- * to for example 1-abc.txt, 2-abc.txt.
- * One of the duplicated files will not have its name changed.
- * Two abc.txt will become 1-abc.txt and abc.txt
- * @param attachments - Array of file objects
- * @returns void. Modifies array in-place.
- */
-export const handleDuplicatesInAttachments = (
-  attachments: IAttachmentInfo[],
-): void => {
-  const names = new Map()
-
-  // fill up the map, the key: filename and value: count
-  attachments.forEach((a) => {
-    if (names.get(a.filename)) {
-      names.set(a.filename, names.get(a.filename) + 1)
-    } else {
-      names.set(a.filename, 1)
-    }
-  })
-
-  // Change names of duplicates
-  attachments.forEach((a) => {
-    if (names.get(a.filename) > 1) {
-      const count = names.get(a.filename) - 1
-      names.set(a.filename, count)
-      a.filename = `${count}-${a.filename}`
-    }
-  })
-}
-
-/**
  * Concatenate response into a string for hashing
  * @param formData Field-value tuples for admin email
  * @param attachments Array of attachments as buffers
